Add files via upload
Task
Complete the getLetter(s) function in the editor. It has one parameter: a string, s, consisting of lowercase English alphabetic letters (i.e., a through z). It must return A, B, C, or D depending on the following criteria:
If the first character in string  is in the {a,i,o,e,u} set , then return A.
If the first character in string  is in the {b,c,d,g,f}set , then return B.
If the first character in string  is in the {h,j,k,l,m} set , then return C.
If the first character in string  is in the {n,p,q,r,s,t,v,w,x,y,z}set , then return D